+++ packages/template-retail-react-app/app/utils/site-utils.js
@@ -1,49 +1,11 @@
 /*
- * Copyright (c) 2021, salesforce.com, inc.
+ * Copyright (c) 2023, Salesforce, Inc.
  * All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
 import {getConfig} from 'pwa-kit-runtime/utils/ssr-config'
-import {getParamsFromPath} from './utils'
-import {absoluteUrl} from './url'
-
-/**
- * This functions takes an url and returns a site object,
- * an error will be thrown if no url is passed in or no site is found
- * @param {string} url
- * @returns {object} site - a site object
- */
-export const resolveSiteFromUrl = (url) => {
-    if (!url) {
-        throw new Error('URL is required to find a site object.')
-    }
-    const {pathname, search} = new URL(absoluteUrl(url))
-    const path = `${pathname}${search}`
-    let site
-
-    // get the site identifier from the url
-    const {siteRef} = getParamsFromPath(path)
-    const sites = getSites()
-
-    // step 1: use the siteRef to look for the site from the sites in the app config
-    // since alias is optional, make sure it is defined before the equality check
-    site = sites.find((site) => site.id === siteRef || (site.alias && site.alias === siteRef))
-    if (site) {
-        return site
-    }
-
-    //Step 2: if step 1 does not work, use the defaultSite value to get the default site
-    site = getDefaultSite()
-    // Step 3: throw an error if site can't be found by any of the above steps
-    if (!site) {
-        throw new Error(
-            "Can't find a matching default site. Please check your sites configuration."
-        )
-    }
-    return site
-}
 
 /**
  * Returns the default site based on the defaultSite value from the app config
