+++ packages/template-retail-react-app/app/utils/site-utils.test.js
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) 2021, salesforce.com, inc.
+ * Copyright (c) 2023, Salesforce, Inc.
  * All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {getDefaultSite, getSites, resolveSiteFromUrl} from './site-utils'
+import {getDefaultSite, getSites} from './site-utils'
+import {resolveSiteFromUrl} from './url' // TODO: Move tests to ./url.test.js
 import {getConfig} from 'pwa-kit-runtime/utils/ssr-config'
 
 import mockConfig from '../../config/mocks/default'
